requirements.txt
asgiref==3.5.2
autopep8==1.7.0
black==22.6.0
click==8.1.3
Django==4.0.6
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
h11==0.13.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.9.1
PyJWT==1.7.1
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.2
whitenoise==6.2.0
